pyproject.toml
[tool.poetry]
name = "psb-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "psb_project"}]

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.15.2"
pandas = "^1.5.2"
psycopg2-binary = "^2.9.5"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
